refractor: rename and seperate files
rename loader.js as model.js
rename ccvLibVolumeShader.js as Shader.js
Move collider-check to its own file
Move entity-collider-check to its own file
